% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utilities.R
\name{G}
\alias{G}
\title{Build transition probability matrix}
\usage{
G(G12, G21)
}
\arguments{
\item{G12}{Probability of jumping to state 2 if currently at state 1.}

\item{G21}{probability of jumping to state 1 if currently in state 2.}
}
\value{
A transition probability matrix
}
\description{
Build transition probability matrix
}
\examples{
G(0.1, 0.2)

}
